+++ lib/.size-snapshot.json
@@ -1,15 +1,15 @@
 {
   "build/dist/material-ui-pickers.esm.js": {
-    "bundled": 116591,
-    "minified": 67636,
-    "gzipped": 15813,
+    "bundled": 112422,
+    "minified": 65859,
+    "gzipped": 15505,
     "treeshaked": {
       "rollup": {
-        "code": 53547,
-        "import_statements": 1273
+        "code": 51948,
+        "import_statements": 1410
       },
       "webpack": {
-        "code": 60685
+        "code": 59038
       }
     }
   },
