+++ lib/src/DateTimePicker/KeyboardDateTimePicker.tsx
@@ -7,14 +7,14 @@ import {
 } from '../_shared/hooks/useKeyboardPickerState';
 import { useUtils } from '../_shared/hooks/useUtils';
 import KeyboardDateInput, { KeyboardDateInputProps } from '../_shared/KeyboardDateInput';
-import { ExtendWrapper2, Wrapper } from '../wrappers/Wrapper';
+import { ExtendWrapper, Wrapper } from '../wrappers/Wrapper';
 import DateTimePicker from './DateTimePicker';
 import DateTimePickerRoot, { BaseDateTimePickerProps } from './DateTimePickerRoot';
 
 export type KeyboardDateTimePickerProps = BaseDateTimePickerProps &
   DateValidationProps &
   BaseKeyboardPickerProps &
-  ExtendWrapper2<KeyboardDateInputProps>;
+  ExtendWrapper<KeyboardDateInputProps>;
 
 export function KeyboardDateTimePicker(props: KeyboardDateTimePickerProps) {
   const {
