+++ lib/src/index.ts
@@ -29,7 +29,7 @@ export { default as DateTimePicker } from './DateTimePicker';
 
 export { default as KeyboardDateTimePicker } from './DateTimePicker/KeyboardDateTimePicker';
 
-export { default as BasePicker } from './_shared/BasePicker';
+export { default as BasePicker } from './typings/BasePicker';
 
 export { default as Calendar } from './DatePicker/components/Calendar';
